A little cleanup and renaming
