+++ Circuitry/IHaveDigitValue.cs
@@ -1,0 +1,8 @@
+﻿namespace Circuitry
+{
+    public interface IHaveDigitValue
+    {
+        public int DisplayValue { get; }
+        public Output Bit0 { get; }
+    }
+}
