+++ Circuitry/SenaryDigit.cs
@@ -4,7 +4,7 @@ namespace Circuitry
 {
     [DebuggerDisplay("Clk={Clk}: {Bit2.State ? 1 : 0}{Bit1.State ? 1 : 0}{Bit0.State ? 1 : 0}")]
 
-    public class SenaryDigit : FourBitDigit
+    public class SenaryDigit : CompoundGate, IHaveDigitValue
     {
         private readonly JKFlipFlop jk1;
         private readonly JKFlipFlop jk2;
@@ -49,7 +49,7 @@ namespace Circuitry
         public Input Clk { get { return clk.Input; } }
         public Output Bit2 { get { return jk3.Q; } }
         public Output Bit1 { get { return jk2.Q; } }
-        public override Output Bit0 { get { return jk1.Q; } }
+        public Output Bit0 { get { return jk1.Q; } }
 
         internal override void RefreshState()
         {
@@ -65,14 +65,14 @@ namespace Circuitry
         {
 			   // TODO: Change to use circuitry
             value %= 6;
-            while (value != DecimalValue)
+            while (value != DisplayValue)
             {
                 Clk.State = !Clk.State;
                 Clk.State = !Clk.State;
             }
         }
 
-        public override int DecimalValue
+        public int DisplayValue
         {
             get
             {
