+++ CircuitryTests/DecadeDigitTests.cs
@@ -12,7 +12,7 @@
             board.Connect(clock.Output, digit.Clk);
 
             // Clk 0: 0000
-            Assert.AreEqual(0, digit.DecimalValue);
+            Assert.AreEqual(0, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -20,7 +20,7 @@
 
             // Clk 1: 0000
             clock.Toggle();
-            Assert.AreEqual(0, digit.DecimalValue);
+            Assert.AreEqual(0, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -28,7 +28,7 @@
 
             // Clk 0: 0001
             clock.Toggle();
-            Assert.AreEqual(1, digit.DecimalValue);
+            Assert.AreEqual(1, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -36,7 +36,7 @@
 
             // Clk 1: 0001
             clock.Toggle();
-            Assert.AreEqual(1, digit.DecimalValue);
+            Assert.AreEqual(1, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -44,7 +44,7 @@
 
             // Clk 0: 0010
             clock.Toggle();
-            Assert.AreEqual(2, digit.DecimalValue);
+            Assert.AreEqual(2, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -52,7 +52,7 @@
 
             // Clk 1: 0010
             clock.Toggle();
-            Assert.AreEqual(2, digit.DecimalValue);
+            Assert.AreEqual(2, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -60,7 +60,7 @@
 
             // Clk 0: 0011
             clock.Toggle();
-            Assert.AreEqual(3, digit.DecimalValue);
+            Assert.AreEqual(3, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -68,7 +68,7 @@
 
             // Clk 1: 0011
             clock.Toggle();
-            Assert.AreEqual(3, digit.DecimalValue);
+            Assert.AreEqual(3, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -76,7 +76,7 @@
 
             // Clk 0: 0100
             clock.Toggle();
-            Assert.AreEqual(4, digit.DecimalValue);
+            Assert.AreEqual(4, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -84,7 +84,7 @@
 
             // Clk 1: 0100
             clock.Toggle();
-            Assert.AreEqual(4, digit.DecimalValue);
+            Assert.AreEqual(4, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -92,7 +92,7 @@
 
             // Clk 0: 0101
             clock.Toggle();
-            Assert.AreEqual(5, digit.DecimalValue);
+            Assert.AreEqual(5, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -100,7 +100,7 @@
 
             // Clk 1: 0101
             clock.Toggle();
-            Assert.AreEqual(5, digit.DecimalValue);
+            Assert.AreEqual(5, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -108,7 +108,7 @@
 
             // Clk 0: 0110
             clock.Toggle();
-            Assert.AreEqual(6, digit.DecimalValue);
+            Assert.AreEqual(6, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -116,7 +116,7 @@
 
             // Clk 1: 0110
             clock.Toggle();
-            Assert.AreEqual(6, digit.DecimalValue);
+            Assert.AreEqual(6, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -124,7 +124,7 @@
 
             // Clk 0: 0111
             clock.Toggle();
-            Assert.AreEqual(7, digit.DecimalValue);
+            Assert.AreEqual(7, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -132,7 +132,7 @@
 
             // Clk 1: 0111
             clock.Toggle();
-            Assert.AreEqual(7, digit.DecimalValue);
+            Assert.AreEqual(7, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsTrue(digit.Bit2.State);
             Assert.IsTrue(digit.Bit1.State);
@@ -140,7 +140,7 @@
 
             // Clk 0: 1000
             clock.Toggle();
-            Assert.AreEqual(8, digit.DecimalValue);
+            Assert.AreEqual(8, digit.DisplayValue);
             Assert.IsTrue(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -148,7 +148,7 @@
 
             // Clk 1: 1000
             clock.Toggle();
-            Assert.AreEqual(8, digit.DecimalValue);
+            Assert.AreEqual(8, digit.DisplayValue);
             Assert.IsTrue(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -156,7 +156,7 @@
 
             // Clk 0: 1001
             clock.Toggle();
-            Assert.AreEqual(9, digit.DecimalValue);
+            Assert.AreEqual(9, digit.DisplayValue);
             Assert.IsTrue(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -164,7 +164,7 @@
 
             // Clk 1: 1001
             clock.Toggle();
-            Assert.AreEqual(9, digit.DecimalValue);
+            Assert.AreEqual(9, digit.DisplayValue);
             Assert.IsTrue(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -172,7 +172,7 @@
 
             // Clk 0: 0000
             clock.Toggle();
-            Assert.AreEqual(0, digit.DecimalValue);
+            Assert.AreEqual(0, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -180,7 +180,7 @@
 
             // Clk 1: 0000
             clock.Toggle();
-            Assert.AreEqual(0, digit.DecimalValue);
+            Assert.AreEqual(0, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -188,7 +188,7 @@
 
             // Clk 0: 0001
             clock.Toggle();
-            Assert.AreEqual(1, digit.DecimalValue);
+            Assert.AreEqual(1, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -196,7 +196,7 @@
 
             // Clk 1: 0001
             clock.Toggle();
-            Assert.AreEqual(1, digit.DecimalValue);
+            Assert.AreEqual(1, digit.DisplayValue);
             Assert.IsFalse(digit.Bit3.State);
             Assert.IsFalse(digit.Bit2.State);
             Assert.IsFalse(digit.Bit1.State);
@@ -214,7 +214,7 @@
             for (int expected = 0; expected < 10; expected++)
             {
                 digit.Set(expected);
-                Assert.AreEqual(expected, digit.DecimalValue);
+                Assert.AreEqual(expected, digit.DisplayValue);
             }
         }
 
@@ -225,10 +225,10 @@
             DecadeDigit digit = new(board);
             ClockGate clock = new();
             board.Connect(clock.Output, digit.Clk);
-            Assert.AreEqual(0, digit.DecimalValue);
+            Assert.AreEqual(0, digit.DisplayValue);
 
             digit.Set(5);
-            Assert.AreEqual(5, digit.DecimalValue);
+            Assert.AreEqual(5, digit.DisplayValue);
