+++ CircuitryTests/Mod3DigitTests.cs
@@ -82,7 +82,7 @@
             for (int expected = 0; expected < 3; expected++)
             {
                 digit.Set(expected);
-                Assert.AreEqual(expected, digit.DecimalValue);
+                Assert.AreEqual(expected, digit.DisplayValue);
             }
         }
     }
