+++ DigitalClockWpf/MainWindow.xaml.cs
@@ -1,14 +1,5 @@
 ﻿using DigitalClockWpf.ViewModel;
-using System.Text;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace DigitalClockWpf
 {
@@ -40,6 +31,11 @@ namespace DigitalClockWpf
         }
 
         private void Stop_Click(object sender, RoutedEventArgs e)
+        {
+            DataContext.Stop();
+        }
+
+        private void Window_Closed(object sender, EventArgs e)
         {
             DataContext.Stop();
         }
