+++ DigitalClockWpf/ViewModel/DigitViewModel.cs
@@ -4,13 +4,13 @@ namespace DigitalClockWpf.ViewModel
 {
     public class DigitViewModel : ViewModelBase
     {
-        private readonly FourBitDigit digit;
+        private readonly IHaveDigitValue digit;
 
         public string Value
         {
             get
             {
-                return digit.DecimalValue.ToString();
+                return digit.DisplayValue.ToString();
             }
         }
 
@@ -18,25 +18,16 @@ namespace DigitalClockWpf.ViewModel
         {
         }
 
-        public DigitViewModel(FourBitDigit digit)
+        public DigitViewModel(IHaveDigitValue digit)
         {
             ArgumentNullException.ThrowIfNull(digit);
             this.digit = digit;
             this.digit.Bit0.StateChanged += Bit_StateChanged;
-            //this.digit.Bit1.StateChanged += Bit_StateChanged;
-            //this.digit.Bit2.StateChanged += Bit_StateChanged;
-            //this.digit.Bit3.StateChanged += Bit_StateChanged;
         }
 
         private void Bit_StateChanged(object? sender, EventArgs e)
         {
             RaisePropertyChanged(() => Value);
         }
-
-        protected override void OnDispose()
-        {
-            //base.OnDispose();
-        }
-
     }
 }
