+++ DigitalClockWpf/ViewModel/MainViewModel.cs
@@ -27,10 +27,5 @@
         {
             clockViewModel.Stop();
         }
-
-        protected override void OnDispose()
-        {
-            //timerViewModel.Dispose();
-        }
     }
 }
