+++ DigitalClockWpf/ViewModel/ViewModelBase.cs
@@ -1,72 +1,124 @@
-﻿namespace DigitalClockWpf.ViewModel
+﻿using System.ComponentModel;
+using System.Diagnostics;
+using System.Linq.Expressions;
+using System.Runtime.CompilerServices;
+
+namespace DigitalClockWpf.ViewModel
 {
     /// <summary>
     /// Abstract base to consolidate common functionality of all ViewModels
     /// </summary>
-    public abstract class ViewModelBase : NotificationObject, IDisposable
+    public abstract class ViewModelBase : INotifyPropertyChanged, INotifyPropertyChanging
     {
-        #region IDisposable Members
-
-        /// <summary>
-        /// Dispose(bool disposing) executes in two distinct scenarios.
-        /// If disposing equals true, the method has been called directly
-        /// or indirectly by a user's code. Managed and unmanaged resources
-        /// can be disposed.
-        /// If disposing equals false, the method has been called by the
-        /// runtime from inside the finalizer and you should not reference
-        /// other objects. Only unmanaged resources can be disposed.
-        /// </summary>
-        /// <param name="disposing"></param>
-        private void Dispose(bool disposing)
+        protected bool SetProperty<T>(ref T storage, T value, [CallerMemberName] string? propertyName = null)
         {
-            // If disposing equals true, dispose all managed
-            // and unmanaged resources.
-            if (disposing)
-            {
-                // Dispose managed resources.
-                OnDispose();
-            }
+            if (Equals(storage, value)) return false;
+
+            RaisePropertyChanging(propertyName);
+            storage = value;
+            RaisePropertyChanged(propertyName);
+            return true;
         }
 
         /// <summary>
-        /// Provided to allow a subclass to dispose of its resources.
+        ///     Occurs when a property value is changing.
         /// </summary>
-        protected virtual void OnDispose()
-        {
-        }
-
-#if DEBUG
-        // Use C# destructor syntax for finalization code.
-        // This destructor will run only if the Dispose method
-        // does not get called.
-        // It gives your base class the opportunity to finalize.
-        // Do not provide destructors in types derived from this class.
-        ~ViewModelBase()
-        {
-            // Do not re-create Dispose clean-up code here.
-            // Calling Dispose(false) is optimal in terms of
-            // readability and maintainability.
-            Dispose(false);
-
-            throw new InvalidOperationException(string.Format("The {0} was not disposed of properly.", ToString()));
-        }
-#endif
+        public event PropertyChangingEventHandler? PropertyChanging;
 
         /// <summary>
-        /// Do not make this method virtual.
-        /// A derived class should not be able to override this method.
+        ///     Raised when a property on this object has a new value.
         /// </summary>
-        public void Dispose()
+        public event PropertyChangedEventHandler? PropertyChanged;
+
+        /// <summary>
+        ///     Provided to allow a subclass to respond to a property changing without connecting to the event handler.
+        /// </summary>
+        protected virtual void OnPropertyChanging(PropertyChangingEventArgs e)
         {
-            // This object will be cleaned up by the Dispose method.
-            // Therefore, you should call GC.SupressFinalize to
-            // take this object off the finalization queue
-            // and prevent finalization code for this object
-            // from executing a second time.
-            Dispose(true);
-            GC.SuppressFinalize(this);
         }
 
-        #endregion IDisposable Members
+        /// <summary>
+        ///     Provided to allow a subclass to respond to a property change without connecting to the event handler.
+        /// </summary>
+        protected virtual void OnPropertyChanged(PropertyChangedEventArgs e)
+        {
+        }
+
+        /// <summary>
+        ///     Raises this object's PropertyChanging event.
+        /// </summary>
+        /// <param name="propertyName">The property that has a new value.</param>
+        //[SuppressMessage("Microsoft.Design", "CA1030:UseEventsWhereAppropriate", Justification = "Method used to raise an event")]
+        protected void RaisePropertyChanging(string? propertyName)
+        {
+            ArgumentException.ThrowIfNullOrWhiteSpace(propertyName);
+            VerifyPropertyName(propertyName);
+
+            var e = new PropertyChangingEventArgs(propertyName);
+            OnPropertyChanging(e);
+
+            PropertyChanging?.Invoke(this, e);
+        }
+
+        /// <summary>
+        ///     Raises this object's PropertyChanging event.
+        /// </summary>
+        /// <typeparam name="T">The type of the property that has a new value</typeparam>
+        /// <param name="propertyExpression">A Lambda expression representing the property that has a new value.</param>
+        protected virtual void RaisePropertyChanging<T>(Expression<Func<T>> propertyExpression)
+        {
+            var propertyName = PropertySupport.ExtractPropertyName(propertyExpression);
+            RaisePropertyChanging(propertyName);
+        }
+
+        /// <summary>
+        ///     Raises this object's PropertyChanged event.
+        /// </summary>
+        /// <param name="propertyName">The property that has a new value.</param>
+        protected void RaisePropertyChanged(string? propertyName)
+        {
+            ArgumentException.ThrowIfNullOrWhiteSpace(propertyName);
+            VerifyPropertyName(propertyName);
+
+            var e = new PropertyChangedEventArgs(propertyName);
+            OnPropertyChanged(e);
+
+            PropertyChanged?.Invoke(this, e);
+        }
+
+        /// <summary>
+        ///     Raises this object's PropertyChanged event.
+        /// </summary>
+        /// <typeparam name="T">The type of the property that has a new value</typeparam>
+        /// <param name="propertyExpression">A Lambda expression representing the property that has a new value.</param>
+        protected virtual void RaisePropertyChanged<T>(Expression<Func<T>> propertyExpression)
+        {
+            var propertyName = PropertySupport.ExtractPropertyName(propertyExpression);
+            RaisePropertyChanged(propertyName);
+        }
+
+        #region Debugging Aides
+
+        /// <summary>
+        ///     Warns the developer if this object does not have
+        ///     a public property with the specified name. This
+        ///     method does not exist in a Release build.
+        /// </summary>
+        [Conditional("DEBUG")]
+        [DebuggerStepThrough]
+        internal void VerifyPropertyName(string propertyName)
+        {
+            if (string.IsNullOrEmpty(propertyName)) return;
+
+            // Verify that the property name matches a real,  
+            // public, instance property on this object.
+            if (TypeDescriptor.GetProperties(this)[propertyName] != null) return;
+
+            var message = "Invalid property name: " + propertyName;
+            Trace.TraceError(message);
+            Debug.Fail(message);
+        }
+
+        #endregion Debugging Aides
     }
 }
